pyproject.toml
[tool.poetry]
name = "2a - account accounting"
version = "0.1.0"
description = "A simple module for the inventory of objects"
authors = ["vocaoo <vocao2308@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
didiator = "^0.3.1"
sqlalchemy = "^2.0.29"
uuid6 = "^2024.1.12"
orjson = "^3.9.15"
structlog = "^24.1.0"
boto3 = "^1.34.70"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
adaptix = "^3.0.0b3"
pillow = "^10.2.0"
di = "^0.79.2"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
exifread = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
